Complete A-Packages
The files were reorganised according to functionality into their
respective packages. Some other additions include:
* Fixing bug with saved files not being decoded properly (Added a decode
  time function)
