+++ docs/README.md
@@ -1,4 +1,4 @@
-# Morpheus User Guide
+# Morpheus.Morpheus User Guide
 
 // Update the title above to match the actual product name
