+++ src/main/java/Morpheus/Commands/Command.java
@@ -1,8 +1,14 @@
+package Morpheus.Commands;
+
+import Morpheus.Utils.Storage;
+import Morpheus.Tasks.Task;
+import Morpheus.Utils.Ui;
+
 import java.util.List;
 
 public abstract class Command {
-    protected String input;
-    protected boolean isExit = false;
+    public String input;
+    public boolean isExit = false;
 
     public Command(String input) {
         this.input = input;
