+++ src/main/java/Morpheus/Commands/DeleteCommand.java
@@ -1,3 +1,9 @@
+package Morpheus.Commands;
+
+import Morpheus.Utils.Storage;
+import Morpheus.Tasks.Task;
+import Morpheus.Utils.Ui;
+
 import java.util.List;
 
 public class DeleteCommand extends Command {
