+++ src/main/java/Morpheus/Tasks/Task.java
@@ -1,3 +1,5 @@
+package Morpheus.Tasks;
+
 public class Task {
     protected String description;
     protected boolean isDone;
