+++ text-ui-test/EXPECTED.TXT
@@ -21,7 +21,7 @@ ___  ___                 _
                   | |
                   |_|
 ------------------------------------------------------------
- Hey there! I'm Morpheus, like the one from The Matrix.
+ Hey there! I'm Morpheus.Morpheus, like the one from The Matrix.
  How can I help you today?
 ------------------------------------------------------------
 ------------------------------------------------------------
